+++ internal/core/data/application/eventpurge.go
@@ -225,7 +225,7 @@ func isReachMaxCap(deviceName, sourceName string, maxCap int64, dic *di.Containe
 	if err != nil {
 		return false, errors.NewCommonEdgeXWrapper(err)
 	}
-	if count < uint32(maxCap) {
+	if count < maxCap {
 		lc.Debugf(
 			"Skip the event retention for the source name `%s` of device `%s`, the event number `%d` is less than the max capacity `%d`.",
 			sourceName, deviceName, count, maxCap,
@@ -252,7 +252,7 @@ func timeBasedEventRetention(deviceName, sourceName string, minCap int64, durati
 			minCap, duration, deviceName, sourceName)
 		// Find the event that the age is within the duration and use offset as minCap to keep data
 		// SELECT * FROM core_data.event WHERE event.origin <= $1 and devicename=$2 and sourcename=$3 ORDER BY origin desc offset $4;
-		event, err := dbClient.LatestEventByDeviceNameAndSourceNameAndAgeAndOffset(deviceName, sourceName, duration.Nanoseconds(), uint32(minCap))
+		event, err := dbClient.LatestEventByDeviceNameAndSourceNameAndAgeAndOffset(deviceName, sourceName, duration.Nanoseconds(), minCap)
 		if errors.Kind(err) == errors.KindEntityDoesNotExist {
 			lc.Debugf("Skip the event retention for the event source '%s' of device '%s', the event number might equal or less than the minCap", sourceName, deviceName)
 			return nil
@@ -286,7 +286,7 @@ func countBasedEventRetention(deviceName, sourceName string, minCap int64, dic *
 	} else {
 		lc.Debugf("MinCap is '%d', purge events by deviceName '%s' and sourceName '%s' to meet the minCap",
 			minCap, deviceName, sourceName)
-		event, err := dbClient.LatestEventByDeviceNameAndSourceNameAndOffset(deviceName, sourceName, uint32(minCap))
+		event, err := dbClient.LatestEventByDeviceNameAndSourceNameAndOffset(deviceName, sourceName, minCap)
 		if errors.Kind(err) == errors.KindEntityDoesNotExist {
 			lc.Debugf("Skip the event retention for the auto event source '%s' of device '%s', the event number might equal or less than the minCap", sourceName, deviceName)
 			return nil
